Fix parsing of large exponent values for NimbleNumberField (#2358)
# Pull Request

## 🤨 Rationale

Fixes #2357 .

## 👩‍💻 Implementation

On `NimbleNumericField`, override `TryParseValueFromString` and
ultimately have it use `double.TryParse` with `NumberStyles.Float |
NumberStyles.Number`.

## 🧪 Testing

Added a `Theory` that changes the text in a `NimbleNumberField` to
various values, and verifies that the string gets parsed into a double
whose `ToString` yields the same string.

Verified that the exponent cases fail without the `NimbleNumberField`
change and pass with it.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.