Update waitForDataUpdatesToRender to work with fit-rows-height token (#2356)
# Pull Request

## 🤨 Rationale

When I went to use the new `ni-nimble-table-fit-rows-height` token in an
Angular app, I realized that the existing `waitForDataUpdatesToRender`
doesn't wait long enough when the table's height is styled with that
token. That's because the new height is calculated, and then the
virtualizer needs to update to render the appropriate number of rows
based on the height. Therefore, I'm updating
`waitForDataUpdatesToRender` on the page object to wait the appropriate
amount of time.

## 👩‍💻 Implementation

Add an additional `await waitForUpdatesAsync()` call within the
`waitForUpdatesToRender` page object function for the Angular table.

## 🧪 Testing

Created unit test for the page object that passes with the change and
fails without it.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>