Disable table row hover styling while scrolling (#2361)
# Pull Request

## 🤨 Rationale

Fixes #2312

## 👩‍💻 Implementation

The TanStack virtualizer exposes an `isScrolling` property. We already
observe the virtualizer for changes (which is triggered when
`isScrolling` changes), so when that is triggered we can update a new
observable property on our (Nimble) Virtualizer class. The table
template uses that property to set/remove an `allow-hover` class on each
row. I updated the row styling to only apply the hover styling when the
`allow-hover` class is present.

I wanted the class to be `scrolling`, but that would require us to use
`:not()` selectors in our styles, which our conventions say to avoid.
Thus I inverted the semantics of the class and picked the affirmative
`allow-hover` (as opposed to `not-scrolling`).

## 🧪 Testing

I don't think there's a practical way to really verifty the end behavior
we want. Our visual tests can't take consistent snapshots while actually
scrolling, and a unit test can't simulaneously scroll and assert the
state. Furthermore, there doesn't seem to be a practical way to trigger
the `:hover` pseudo-class in a unit test, so we can't test that our CSS
is written properly. The best I could do was to mock the virtualizer's
`isScrolling` property to return `true` and test that the `allow-hover`
class does not get applied.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.