Add icons for install and uninstall software (#2383)
# Pull Request

## 🤨 Rationale

Resolves #2354.

## 👩‍💻 Implementation

Followed the new icon instructions in CONTRIBUTING. I used the names and
synonyms suggested on the issue but changed 'box' to 'rectangle' to
align better with terminology already in use.

Also fixed the ordered list numbering in the nimble-tokens CONTRIBUTING.

## 🧪 Testing

Locally verified the icons can be colored by severity and that they
appear in Storybook and the Chromatic diff.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.