Fix two sources of references to detached rows (#2392)
# Pull Request

## 🤨 Rationale

While investigating a [memory leak
issue](https://ni.visualstudio.com/DevCentral/_workitems/edit/2843309)
reported in AzDO, we found a couple places in the Nimble table code that
wasn't cleaning up observers/notifiers, causing us to hold references to
disconnected row elements.

There is also a FAST bug that contributes to the leak (which is [getting
fixed](https://github.com/microsoft/fast/pull/7023)), so the leak will
still not be fixed until this PR and the FAST one are both in.

## 👩‍💻 Implementation

1. Rows observe columns, which results in there being a reference chain
from each column to each row. When the row is detached, we must remove
the column observers.
2. The keyboard navigation manager has logic for storing the row
notifiers it creates, so that it can clean them up at the appropriate
time. However, by a simple oversight, we were never actually adding the
notifiers to the array.

## 🧪 Testing

Verified (in Chrome dev tools) that after these changes, detached rows
are no longer referenced via the columns or keyboard navigation manager.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.