Fix tabs toolbar content alignment (#2451)
# Pull Request

## 🤨 Rationale

[SLE bug: Nimble | nimble-tabs-toolbar | alignment of the child elements
has been
altered.](https://dev.azure.com/ni/DevCentral/_workitems/edit/2913486)

## 👩‍💻 Implementation

Just added styling that will expand the `end` slot to fill the rest of
the space of the `Tabs` width.

## 🧪 Testing

There is a Chromatic test in place that is already checking this case,
but I must have glossed over that change when examining it for the
scrollable tabs submission.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.