Pin FAST foundation and element versions until new versions vetted (#2455)
# Pull Request

## 🤨 Rationale

Unit tests on Firefox seem to be sporadically timing out when we update
to the latest exports of `fast-foundation` (2.50.0) and `fast-element`
(1.14.0). We want to pin our dependency versions to the prior versions
until we are confident that the new versions do not introduce real
problems for clients.

## 👩‍💻 Implementation

Pinned `fast-element` to 1.13.0 and `fast-foundation` to 2.49.6 in the
three packages that reference them.

## 🧪 Testing

N/A

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.