Fix issue with tab panel not sizing to height of tabs component. (#2458)
# Pull Request

## 🤨 Rationale

The update to the `Tabs` component introducing scroll buttons #2440 also
updated the `display` setting to `flex`. This resulted in a missed
change in behavior with how the `TabPanel` might size to the height of
the `Tabs` component. See [this
discussion](https://dev.azure.com/ni/DevCentral/_git/Skyline/pullrequest/821815?discussionId=7029412).

## 👩‍💻 Implementation

Just needed to set a style on `[part="tabpanel"]` to have `flex: 1;`.

## 🧪 Testing

Added a Chromatic test that would fail without the style change.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.