Checkbox error state HLD update (#2467)
# Pull Request

## 🤨 Rationale

During development of #2018, I realized that implementing the error
state of the checkbox the way it was spec'ed out would be much more
difficult than originally thought because making only the label and
square checked indicator clickable would break a number of tests (inside
& outside of nimble) and would be against ARIA guidelines. Therefore, we
decided to keep the effort of the feature more in line with what our
original estimates were by deferring the work to update which portion of
the control was clickable.

## 👩‍💻 Implementation

Update HLD

## 🧪 Testing

None

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.