Upgrade Storybook to 8.4 (#2470)
# Pull Request

## 🤨 Rationale

I saw that Storybook released a new minor version which is supposed to
have a bunch of improvements. An exciting sounding one: "We cut bundle
size in half" though the package-lock has about as many additions as
subtractions


## 👩‍💻 Implementation

Followed [their docs](https://storybook.js.org/blog/storybook-8-4/)
which basically involved:
1. Run `npx storybook@latest upgrade`
2. Agree to the prompts to auto-update stuff
3. The tool also suggested we might be able to remove our React
dependency but I didn't try this because I know we create some custom
React components so it seems correct for us to depend on it

## 🧪 Testing

1. Manually inspected our stories including the ones that use custom
React components
2. verified Chromatic did a full rebuild, no turbosnaps, and there are
no diffs

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>