Add error state to nimble-checkbox (#2478)
# Pull Request

## 🤨 Rationale

This is part of #2018 and also resolves #2091

## 👩‍💻 Implementation

The `nimble-checkbox` now implements `ErrorPattern` to add `error-text`
and `error-visible` to its API.

The bulk of the implementation is updating the template and styles of
the checkbox. These changes include:
- Adding error icon/text to the template with appropriate styles
- Fix the wrapping behavior to keep the checkbox aligned with the top
line of text rather than centered within the control
- Update the checkbox to have a default min-height of `controlHeight`
(see #2091)

## 🧪 Testing

- Existing unit tests pass
- Manually tested in storybook
- Extended matrix tests to cover more cases -- error states and wrapping
text

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.