Error text overflow (#2432)
# Pull Request

## 🤨 Rationale

Fixes #2431 

## 👩‍💻 Implementation

Updated the error pattern to be a mixin rather than an interface so that
controls using the error pattern don't have to redefine the
properties/attributes that the pattern uses. Additionally, logic has
been added to the mixin (beyond what the interface supported) to use the
`overflow` behavior to add a title to the error text when it overflows.

## 🧪 Testing

Manually tested in storybook that all updated components:
- don't have a `title` on the error text when the error text is short
- have a `title` on the error text when the error text overflows
- the overflow on the error text does not affect any existing overflow
behavior on the component

New unit tests for the error pattern mixin

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>