Fork number-field and text-field templates (#2495)
# Pull Request

## 🤨 Rationale

Fork the FAST templates for the number-field and the text-field in
preparation for adding a required-visible icon to the label (see #2100)

## 👩‍💻 Implementation

Forked the FAST templates for the number-field and the text-field with
no functional changes.

## 🧪 Testing

Verified tests still pass and no visual changes were introduced to those
components.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>