Required indicator on radio-group, select, combobox, and text-area (#2496)
# Pull Request

## 🤨 Rationale

Add `required-visible` attributes to the radio-group, select, combobox,
and text-area components as part of #2100.

There is still a need to add this attribute to the text-field and
number-field. That will be done once the templates for those components
have been forked (#2495).

## 👩‍💻 Implementation

Within the `patterns` directory, I created a mixin that can be used on
elements that should have a `requiredVisible` property and
`required-visible` attribute. Additionally, shared styling was added for
the `requiredVisible` state and a function was created to generate the
template for an element's label. This is a function rather than a
constant template because each element already had an existing label
structure that I didn't want to change as part of this PR.

The template of each element included in this PR was also updated to set
`aria-required` based on the value of `requiredVisible`.

I also updated the styling of icons to solve a problem I encountered
when shrinking the asterisk icon down to 5x5.

## 🧪 Testing

- Manually tested
- New matrix tests
- Unit tests for the new mixin
- New unit tests for the ARIA behavior of each element

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Fred Visser <1458528+fredvisser@users.noreply.github.com>