Replace deprectated `importsNotUsedAsValues` option with `verbatimModuleSyntax` (#2512)
# Pull Request

## 🤨 Rationale

The
[`importsNotUsedAsValues`](https://www.typescriptlang.org/tsconfig/#importsNotUsedAsValues)
TypeScript option is deprectated and will stop working in TypeScript
5.5. It is replaced by
[`verbatimModuleSyntax`](https://www.typescriptlang.org/tsconfig/#verbatimModuleSyntax).

My motivation for this change is that while updating Nimble to Angular
18, I found that `importsNotUsedAsValues` was forcing me to do a
type-only import of `Router` and `ActivatedRoute` [in
`router_link.ts`](https://github.com/ni/nimble/blob/a7927039f1da96ac4672c5550d6b8f55071f125d/packages/angular-workspace/nimble-angular/src/thirdparty/directives/router_link.ts#L39).
But doing so caused a different error for
`NimbleAnchorRouterLinkWithHrefDirective`:

```
error NG2016: The directive NimbleAnchorRouterLinkWithHrefDirective inherits its constructor from RouterLink, but the latter has a constructor parameter that is not compatible with dependency injection. Either add an explicit constructor to NimbleAnchorRouterLinkWithHrefDirective or change RouterLink's constructor to use parameters that are valid for DI.
```

Both `Router` and `ActivatedRoute` are parameters to the `RouterLink`
constructor. The problem is that if we do a type-only import of `Router`
and `ActivatedRoute`, the compiler doesn't know that they are decorated
to be injectable. Using `verbatimModuleSyntax` solves this by allowing
us to import from the definition file rather than the type declaration
file for those types, even though we're not directly using the
definition.

This PR changes the option for the Angular projects, as well as
`nimble-components`, `spright-components`, and `storybook`.

## 👩‍💻 Implementation

Had to update all imports of type symbols to explicitly say `type`.

## 🧪 Testing

Builds and existing tests pass.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>