Update to latest .net sdk 8.0.405 (#2518)
# Pull Request

## 🤨 Rationale

Builds on main started failing with messages like:

>
/home/runner/work/nimble/nimble/packages/blazor-workspace/Examples/Demo.Client/Demo.Client.csproj
: error NU1004: The package reference Microsoft.NET.Sdk.WebAssembly.Pack
version has changed from [8.0.11, ) to [8.0.12, ).The packages lock file
is inconsistent with the project dependencies so restore can't be run in
locked mode. Disable the RestoreLockedMode MSBuild property or pass an
explicit --force-evaluate option to run restore to update the lock file.
[/home/runner/work/nimble/nimble/packages/blazor-workspace/BlazorWorkspace.sln]


But if you check for references to 8.0.12 assemblies you won't find
them. I think what is happening is that the GitHub agents have the
latest v8 dotnet installed (v8.0.405) which supersedes the older
(v8.0.404) we try to install. So the newer assemblies (v8.0.12) seem to
completely replace what the lockfile targets (v8.0.11).

You can reproduce the above errors by just having the newer dotnet sdk
(v8.0.405) installed and running a nimble build prior to this PR.

## 👩‍💻 Implementation

Updated to the latest dotnet sdk and ran `npm run
update-lock-filesupdate-lock-files`.

## 🧪 Testing

Rely on CI.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.