Allow an item in the toolbar to lose focus by clicking the toolbar itself (#2521)
# Pull Request

## 🤨 Rationale

Resolves #2494

## 👩‍💻 Implementation

Removed `delgatesFocus: true` on the toolbar so that clicking on an
empty space within the toolbar will cause the currently focused item to
lose focus.

Note: The use case described in the referenced issue may still be a bit
problematic given all the editable fields (date pickers, numerics, etc)
because the user may expect to use the left/right arrow keys to navigate
within that component, but the toolbar will handle the key event first
and move focus to a different item in the keyboard. That behavior is not
affected in any way by this change.

## 🧪 Testing

Manually tested with a numeric field and text field in the toolbar.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>