Fix for Select page object clickOption method. (#2527)
# Pull Request

## 🤨 Rationale

When refactoring some tests for the notebook selector in SLE, I ran into
some issues when trying to use the `clickOption` method on the page
object for a dynamic filtering scenario. In this specific scenario,
using the index-based click method was needed over
`clickOptionWithDisplayText`, as the use-case had two options with the
same display text (under different groups).