Add usage guidance about table column mapping performance (#2548)
# Pull Request

## 🤨 Rationale

The impetus for #2531 was performance issues due to creating a mapping
element for each row of table data. We should explicitly suggest not
doing this.

## 👩‍💻 Implementation

Added usage guidance to the table column mapping story.

## 🧪 Testing

Inspect storybook build

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>