Update table number text column docs for unit to use slots API category (#2587)
# Pull Request

## 🤨 Rationale

I noticed that the docs for how to set units on
`nimble-table-column-number-text` were incorrect. They correctly state
that you specify units by slotting a unit element in the default slot.
But those docs appeared in the attributes section of the API table.

## 👩‍💻 Implementation

Change the API category to "slots" and the display name to "default"

## 🧪 Testing

Local storybook

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.