[HLD] Readonly appearance for disabled controls (#2592)
# Pull Request

## 🤨 Rationale

There is a desire to have read-only controls in forms, but this is not
supported out-of-the box by forms or by all nimble controls. This HLD
proposes a way to support this capability.

## 👩‍💻 Implementation

N/A

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>