Rename matrix tests (#2598)
# Pull Request

## 🤨 Rationale

This is a follow up change from #2596 to rename the matrix tests to use
`$` as a delimiter when the name includes which dimension a test is
using.

## 👩‍💻 Implementation

Rename matrix tests using the following guidelines:
- Don't include the name of the component in the test name
- If the theme is part of the name of the test, put it first without the
background color (e.g. without `WhiteBackground`)
- Don't include theme in the name if it isn't part of the test
dimension. For example, if a test is run only for one theme to test
something unrelated to theme, it doesn't need to be included in the
name.
- Use `$` as a delimiter in the test name when the name includes
multiple constant dimension values

In addition, I split the test for a hidden radio item into its own test
file under `Tests/Radio`. It had been a story under `Tests/Radio Group`.

## 🧪 Testing

Tested in storybook locally

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>
Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>