Add appearance-readonly to nimble-number-field (#2600)
# Pull Request

## 🤨 Rationale

Add `appearance-readonly` attribute to the `nimble-number-field` to
style the component as readonly when it is disabled.

I also realized that the cursor was wrong for the `disabled +
appearanceReadOnly` state on the `nimble-text-field`, so I fixed it.

## 👩‍💻 Implementation

- Add `appearance-readonly` attribute (tied to `appearanceReadOnly`
property) on the text field
- Updated styles to make `disabled + appearanceReadOnly` look like the
`readonly` styles
- Updated storybook docs
- Updated matrix tests
- As part of the matrix test updates, I split the number field matrix
tests into 3 different tests (one for each theme) because the canvas of
the story got too tall for chromatic otherwise.

## 🧪 Testing

- Manually tested
- Chromatic test

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.