Update nimble-number-field to support frameless and full-bleed styling (#2606)
# Pull Request

## 🤨 Rationale

Updated the `nimble-number-field` appearance modes to include
`frameless` and to also have the option to be `full-bleed` (which only
applies when the appearance is `frameless`). This is part of the effort
to be able to render more nimble controls in a "read-only" manner where
the value is aligned with the label.

See #2592 

## 👩‍💻 Implementation

- Add `frameless` appearance for the number field
- Add `full-bleed` attribute to the number field

## 🧪 Testing

- Manual testing in storybook
- Updated chromatic tests
- Because the theme tests got too big for the number field, I split each
story into 2 -- one with the step buttons visible and one with the step
buttons hidden.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>