Add appearance-readonly to nimble-select and nimble-combobox (#2602)
# Pull Request

## 🤨 Rationale

Add `appearance-readonly` attribute to the `nimble-select` and
`nimble-combobox` to style the components as readonly when they are
disabled.

## 👩‍💻 Implementation

- Add `appearance-readonly` attribute (tied to `appearanceReadOnly`
property) on the select and combobox
- Added styles to make `disabled + appearanceReadOnly` look like other
readonly controls (neither select nor combobox has a `readonly` property
or styles today)
- Updated storybook docs
- Updated matrix tests
- As part of the matrix test updates, I split the select matrix story
into 3 stories (one for each theme) because chromatic was timing out
loading the story

## 🧪 Testing

- Manually tested
- Chromatic test

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.