Update select and combobox to support frameless and full-bleed styling (#2613)
# Pull Request

## 🤨 Rationale

Updated the `nimble-select` and `nimble-combobox` appearance modes to
include `frameless` and to also have the option to be `full-bleed`
(which only applies when the appearance is `frameless`). This is part of
the effort to be able to render more nimble controls in a "read-only"
manner where the value is aligned with the label.

See #2592 

## 👩‍💻 Implementation

- Add `frameless` appearance for the drop-down appearance modes
- Add `full-bleed` attribute to the select and combobox

## 🧪 Testing

- Manual testing in storybook
- Updated chromatic tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.