Add icon metaphors (#2620)
# Pull Request

## 🤨 Rationale

Some of the nimble icons will be used to represent actions for workflows
on test plans. Since these icons have been approved for this usage, I'm
ensuring the alias of the icon within the workflows code is represented
as a metaphor within the nimble documentation.

Most of these already existed as metaphors, but there were a few that
needed to be added.

## 👩‍💻 Implementation

N/A

## 🧪 Testing

Verified the storybook story still renders as expected.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.