Add end slot for tabs toolbar (#2637)
# Pull Request

## 🤨 Rationale

#2635: Tab toolbar - Support removing splitter when Toolbar buttons are
right aligned

## 👩‍💻 Implementation

I'm proposing an automatic behavior for showing/hiding the separator
based on whether any content has been placed in the default slot. In
order to hide the separator, _all_ content must be placed in the `end`
slot. This results in a non-breaking change for current usages.

In order to get the desired behavior listed in the linked work item,
clients will have to update their right-aligned elements to be placed
into the `end` slot (and ideally remove the now-redundant positioning
styling).

## 🧪 Testing

Added storybook matrix tests.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>