Localize decimal separator in number-field (#2641)
# Pull Request

## 🤨 Rationale

Adding support for localization of the decimal separator in the
`nimble-number-field`, similar to how the number-text table column
localizes the separator character. Part of #1855.

## 👩‍💻 Implementation

- As proposed in
[HLD](https://github.com/ni/nimble/blob/main/packages/nimble-components/src/number-field/specs/localized-decimal-separator-hld.md),
overriding several functions from the `@ni/fast-components` number-field
implementation to support dynamically updating the view and input filter
based on the value of the `lang` design token. `Intl.NumberFormat` is
used to determine the correct decimal separator character for the given
locale identifier (i.e. value of `lang`).
- Changed `inputmode` attribute of HTML `input` in the number-field's
shadow DOM from `"numeric"` to `"text"`. Manually testing showed the
desired input mode of
[`decimal`](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Global_attributes/inputmode#decimal)
did not include negative in the virtual keyboard so instead a full
keyboard was chosen.
- Updated documentation, including changing the type of the `value`
property to string (from number, which was confusing). I also added
documentation for the `valueAsNumber` property, which helps reinforce
the message that `value` is a string.

## 🧪 Testing

Added automated unit tests to verify the new functionality.
Additionally, since the added function overrides contain logic not
directly related to this new functionality (i.e. the bits of logic that
needed to be duplicated from the base implementation), I added a couple
tests to ensure we haven't introduced regressions.

I added a (minimal) page object for the number-field as well.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>