Spright `ChatInput` remove `resetInput` (#2666)
# Pull Request

## 🤨 Rationale

The design for the chat input calls to _always_ clear the text in the
input once a user sends the message. There does not seem to be a
compelling enough reason to make the client call a method in response to
the `send` event to make this happen ([additional
context](https://github.com/ni/nimble/pull/2611#discussion_r2226714997)).

## 👩‍💻 Implementation

Make `resetInput` private and call it from the `sendButtonClickHandler`.

## 🧪 Testing

Unit test.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.