Add theme aware token for calendar resource background conflict color (#2665)
# Pull Request

## 🤨 Rationale

Add theme aware token for calendar resource background conflict color
based on the mockup.
Mockup link:
[https://www.figma.com/design/CkYDC25GDGsxQFfqvv1C8T/Test-Plan-automation?node-id=1998-18796&t=iuJckVnLRQHy6gGw-0](https://www.figma.com/design/CkYDC25GDGsxQFfqvv1C8T/Test-Plan-automation?node-id=1998-18796&t=iuJckVnLRQHy6gGw-0
)
## 👩‍💻 Implementation

- A test plan conflict is nothing but two or more test plans gets
scheduled in the common time slot on the same resource. For indicating
the test plan conflict to the user a red background will be shown in the
conflicting area of the resource. For this functionality we have added
`calendarRowBackgroundConflictColor` token with respect to the mockup
for all three themes.

## 🧪 Testing

- Update the existing calendar story to add the new tokens and manually
validate the colors by inspecting.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Signed-off-by: gokulprasanth-ni <gokulprasanth.ravi@emerson.com>
Signed-off-by: gokulprasanth-ni <131153319+gokulprasanth-ni@users.noreply.github.com>
Co-authored-by: gokulprasanth-ni <gokulprasanth.ravi@emerson.com>