Add background api for chat conversation (#2694)
# Pull Request

## 🤨 Rationale

We have use cases for the chat UI that require the background to be
transparent in order for a larger app background to appear through the
chat conversation area.

## 👩‍💻 Implementation

Add a simple boolean attribute to control whether or not to show the
conversation background or not.

## 🧪 Testing

Chromatic tests.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>